requirements.txt
# requirements.txt
py-algorand-sdk>=2.10.0,<3.0.0
pyteal>=0.27.0,<1.0.0
python-dotenv>=1.1.1,<2.0.0
streamlit>=1.49.0,<2.0.0
uvloop>=0.21.0,<1.0.0; sys_platform != 'win32'
qrcode>=8.2,<9.0.0
Pillow>=11.3.0,<12.0.0
reportlab>=4.4.3,<5.0.0
ruff>=0.12.11,<1.0.0
black>=25.1.0,<26.0.0
mypy>=1.17.1,<2.0.0
sqlfluff>=3.4.2,<4.0.0
detect-secrets>=1.5.0,<2.0.0
pre-commit>=4.3.0,<5.0.0
